requirements.txt
alabaster
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
astroid
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8
Babel
backcall==0.2.0
bcrypt
black==19.10b0
bleach==3.3.0
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
chardet
click
cloudpickle
colorama==0.4.4
cryptography
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch
docutils
entrypoints==0.3
flake8
future
hpsklearn
hyperopt==0.2.5
imagesize
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata
intervaltree
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort
jedi==0.18.0
Jinja2==2.11.3
joblib
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring
kiwisolver==1.3.1
lazy-object-proxy
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mne==0.22.0
mypy-extensions==0.4.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nose==1.3.7
notebook==6.2.0
numpy
numpydoc
olefile==0.46
packaging==20.9
pandas
pandocfilters==1.4.3
paramiko
parso==0.8.1
pathspec==0.7.0
pexpect
pickleshare==0.7.5
Pillow
pluggy
prometheus-client==0.9.0
prompt-toolkit==3.0.17
psutil
ptyprocess
pyaml==20.4.0
pycodestyle
pycparser==2.20
pydocstyle
pyflakes
Pygments==2.8.1
pylint
pyls-black
pyls-spyder
PyNaCl
pyOpenSSL
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks
python-dateutil==2.8.1
python-jsonrpc-server
python-language-server
PyYAML==5.4.1
pyzmq==22.0.3
QDarkStyle==2.8.1
QtAwesome
qtconsole
QtPy==1.9.0
regex
requests
rope
Rtree
scikit-learn
scikit-optimize @ git+https://github.com/scikit-optimize/scikit-optimize.git@530da127c0e3d92fc5018115585e73fecced12a5
scipy
Send2Trash==1.5.0
six==1.15.0
snowballstemmer
sortedcontainers
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder-kernels
terminado==0.9.3
testpath==0.4.4
textdistance
threadpoolctl
three-merge
toml
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typed-ast
typing-extensions
ujson
urllib3
watchdog
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton
wincertstore==0.2
wrapt
yapf
zipp
alabaster
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
astroid
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8
Babel
backcall==0.2.0
bcrypt
black==19.10b0
bleach==3.3.0
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
chardet
click
cloudpickle
colorama==0.4.4
cryptography
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch
docutils
entrypoints==0.3
flake8
future
hpsklearn
hyperopt==0.2.5
idna
imagesize
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata
intervaltree
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort
jedi==0.18.0
Jinja2==2.11.3
joblib
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring
kiwisolver==1.3.1
lazy-object-proxy
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mne==0.22.0
mypy-extensions==0.4.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nose==1.3.7
notebook==6.2.0
numpy
numpydoc
olefile==0.46
packaging==20.9
pandas
pandocfilters==1.4.3
paramiko
parso==0.8.1
pathspec==0.7.0
pexpect
pickleshare==0.7.5
Pillow
pluggy
prometheus-client==0.9.0
prompt-toolkit==3.0.17
psutil
ptyprocess
pyaml==20.4.0
pycodestyle
pycparser==2.20
pydocstyle
pyflakes
Pygments==2.8.1
pylint
pyls-black
pyls-spyder
PyNaCl
pyOpenSSL
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks
python-dateutil==2.8.1
python-jsonrpc-server
python-language-server
PyYAML==5.4.1
pyzmq==22.0.3
QDarkStyle==2.8.1
QtAwesome
qtconsole
QtPy==1.9.0
regex
requests
rope
Rtree
scikit-learn
scikit-optimize @ git+https://github.com/scikit-optimize/scikit-optimize.git@530da127c0e3d92fc5018115585e73fecced12a5
scipy
Send2Trash==1.5.0
six==1.15.0
snowballstemmer
sortedcontainers
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder-kernels
terminado==0.9.3
testpath==0.4.4
textdistance
threadpoolctl
three-merge
toml
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typed-ast
typing-extensions
ujson
urllib3
watchdog
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton
wincertstore==0.2
wrapt
yapf
zipp
keras
